+++ src/components/_auth/SignInSteps/TermsStep.styles.ts
