+++ src/components/_auth/SignInSteps/TermsStep.tsx
@@ -8,7 +8,7 @@ import { transitions } from '@/styles'
 import { StepFooter, StepWrapper } from './SignInSteps.styles'
 import { ContinueButton, ScrollButton, TermsBox, TermsOverlay, TextWrapper } from './TermsStep.styles'
 
-import { TermsOfService } from '../TermsOfService'
+import { TermsOfService } from '../../TermsOfService'
 
 export const TermsStep: React.FC = () => {
   const [hasScrolledToBottom, setHasScrolledToBottom] = useState(false)
